requirements.txt
pygame
numpy
matplotlib
jupyterlab
pillow 
imageio 
imageio-ffmpeg